The PubSub compiled without error
